+++ foundationdb/tests/api.rs
@@ -1,4 +1,4 @@
-use foundationdb::fdb_api::FdbApiBuilder;
+use foundationdb::api::FdbApiBuilder;
 use std::thread;
 
 #[test]
